second instance of float to int needed

+++ daymetpy/daymetpy.py
@@ -119,7 +119,7 @@ def download_Daymet(site="Daymet",lat=36.0133,lon=-84.2625,start_yr=1980,end_yr=
         if as_dataframe:
             import pandas as pd
             df = pd.read_csv(daymet_file, header=6)
-            df.index = pd.to_datetime(df.year.astype(str) + '-' + df.yday.astype(str), format="%Y-%j")
+            df.index = pd.to_datetime(df.year.astype(int).astype(str) + '-' + df.yday.astype(int).astype(str), format="%Y-%j")
             df.columns = [c[:c.index('(')].strip() if '(' in c else c for c in df.columns ]
             return df
         else:
